Use custom icon for `lib` directory in the Project Tree
