+++ gradle.properties
@@ -42,7 +42,7 @@ intellij.update.since.until.build=true
 
 intellij.plugin.id=com.intellij.idea.plugin.sap.commerce
 intellij.plugin.name=SAP Commerce Developers Toolset
-intellij.plugin.version=2025.2.3
+intellij.plugin.version=2025.2.4
 intellij.plugin.since.build=252.23892.409
 intellij.plugin.until.build=252.*
