+++ modules/project/ui/src/sap/commerce/toolset/project/view/HybrisProjectViewNodeDecorator.kt
@@ -15,6 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
+
 package sap.commerce.toolset.project.view
 
 import com.intellij.ide.projectView.PresentationData
