+++ modules/project/ui/src/sap/commerce/toolset/project/view/nodes/YProjectViewModuleGroupNode.kt
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package sap.commerce.toolset.project.view
+package sap.commerce.toolset.project.view.nodes
 
 import com.intellij.ide.projectView.PresentationData
 import com.intellij.ide.projectView.ViewSettings
